+++ src/tools/singlejar/output_jar.cc
@@ -235,7 +235,7 @@ int OutputJar::Doit(Options *options) {
   }
 
   // Then copy source files' contents.
-  for (int ix = 0; ix < options_->input_jars.size(); ++ix) {
+  for (size_t ix = 0; ix < options_->input_jars.size(); ++ix) {
     if (!AddJar(ix)) {
       exit(1);
     }
@@ -646,7 +646,7 @@ void OutputJar::AppendToDirectoryBuffer(const CDH *cdh, off_t lh_pos,
   const Zip64ExtraField *zip64_ef = cdh->zip64_extra_field();
   const int zip64_attr_count = zip64_ef == nullptr ? 0 : zip64_ef->attr_count();
   const bool lh_pos_needs64 = ziph::zfield_needs_ext64(lh_pos);
-  size_t out_zip64_attr_count;
+  int out_zip64_attr_count;
   if (zip64_attr_count > 0) {
     out_zip64_attr_count = zip64_attr_count;
     // The number of attributes may remain the same, or it may increase or
@@ -881,7 +881,7 @@ ssize_t OutputJar::AppendFile(int in_fd, off_t offset, size_t count) {
   }
   ssize_t total_written = 0;
 
-  while (total_written < count) {
+  while (static_cast<size_t>(total_written) < count) {
     size_t len = std::min(kBufferSize, count - total_written);
     ssize_t n_read = pread(in_fd, buffer.get(), len, offset + total_written);
     if (n_read > 0) {
